Use a default 'General' category when creating a new Site Template.
In order for a Site Template to appear in the frontend picker, the template
must be associated with one or more categories linked to the given
group type. To minimize confusion, we install a default category 'General'
which is linked to all group types, and check this category by default
when creating a new template.

See cuny-academic-commons/commons-in-a-box#398.