Skip the 'astra_partials_config_cache' option when copying site templates.
This option, created by the Astra theme, contains references to classes
that are defined by Astra. As such, those classes do not exist when
the template options are copied to the new site. When CBOX unserializes
the option in order to do a URL swapout, those missing classes trigger
a fatal error on PHP 8+.

See cuny-academic-commons/commons-in-a-box#433.