Yet another overhaul of the flow of default navigation items through the template system.
Several primary changes.

First, because template creation takes place on the primary site, we must offload
the generation of default navigation items for template sites onto separate processes.
This ensures that they are set up in the proper theme context, meaning that
block themes will be properly recognized, and classic menu areas will be pulled from
the appropriate theme (ie not openlab-theme).

Second, we introduce a new navigation-link block variation for 'Group Home Link'.
This block, which is intended to be used only in templates, provides the necessary
placeholder so that the necessary swap-outs can take place when cloning from the
template site.

See cuny-academic-commons/commons-in-a-box#511.